Fit bipartite network model

Fit bipartite network model

Usage

fit_model(
  data = NULL,
  plant_effort = c("param", "data"),
  animal_pref = c("fixed", "varying"),
  plant_abun = c("param", "data"),
  animal_abun = c("param", "data"),
  beta = 0.01,
  alpha_sigma = 1,
  alpha_tau = 1,
  ...
)

Arguments

data

A named list containing the required data, as obtained from prepare_data().

plant_effort

How to model the sampling effort allocated to each plant. If unknown, "param" (default) estimates a single parameter during the modelling step that applies to all plants, \(C\). Alternatively, "data" allows measured values of sampling effort to be supplied in data, \(C_i\).

animal_pref

How to model the increase in visits of animals to plants with which they share a link relative to unlinked plants. "fixed" (default), fits a single preference parameter for all animals during the modelling step, \(r\). "varying" fits a separate preference parameter for each animal type, \(r_i\).

plant_abun

How the model treats the abundances of interacting plants. "param" (default) treats these abundances as a parameter that the model estimates, \(\sigma_i\). "data" treats these abundances as measured and requires the user to supply the absolute abundances of each plant in data.

animal_abun

How the model treats the abundances of interacting animals. "param" (default) treats these abundances as a parameter that the model estimates, \(\tau_i\). "data" treats these abundances as measured and requires the user to supply the absolute abundances of each animal in data.

beta

Rate parameter of exponential prior on the preference parameter or parameters \(r\) or \(r_i\). The default, 0.01, is suitable for small counts and should be increased for larger counts. Use plot_prior() to visualize.

alpha_sigma

Vector of concentration parameters for Dirichlet prior on the relative abundances of interacting plants, \(\sigma_i\), when plant_abun = "param". Ignored otherwise. Defaults to 1.

alpha_tau

Concentration parameters of Dirichlet prior on the relative abundances of interacting plants, \(\tau_i\), when animal_abun = "param". Ignored otherwise. Defaults to 1.

...

Further arguments passed to cmdstanr::sample(), like iter_warmup, iter_sampling, or thin, among others.

Value

A fitted model (cmdstanr::CmdStanMCMC() object).

Details

This function allows fits generalized variations of the bipartite network structure model of Young et al. 2021 doi:10.1038/s41467-021-24149-x. Briefly, the observed interaction count between each plant and animal pair is treated as a random draw from a Poisson distribution with a mean of: $$\mu_{ij} = C_i\sigma_i\tau_j\left(1 + r_iB_{ij}\right).$$ Here, the subscripts \(i\) and \(j\) are indices that correspond to the plant (row) and animal (column) species, respectively. \(C_i\) is the observer's sampling effort allocated to each plant species; we expect to see more interactions with plants the longer they are observed. \(\sigma_i\) and \(\tau_i\) are the relative contributions of each plant and animal to the total pool of interactions. We expect to observe higher interaction counts for species that contribute most to the interactions made by their guild. The observed interaction count should be compounded when species of both guilds are those that are most interactive. Importantly, these terms capture more than just each species' relative abundances in the their guild. An animal that is highly active but numerically rare animal could have a higher \(\tau_i\) than a numerically dominant, but less active animal. \(r_i\) is the increase in visits of animals to plants with which they share a link relative to unlinked plants, and \(B_{ij}\) is the inferred absence (0) or presence (1) of a link between the plant-animal pair.

Prior distributions

The default priors are very weak to accommodate a wide range of empirical datasets. They may fail for exceptional datasets and they do not capture system-specific prior information that users may have from additional data.

The animal preference parameter, \(r_i\), has an exponential prior with a rate parameter beta. The default, \(\mathrm{exp}(0.01)\), implies a mean of 1/0.01 = 100 with 95% of the probability mass falling between about 2.5-370. Increase beta for animal communities that weakly increase visitation in the presence of a link, or decrease for communities that strongly increase visitation in the presence of a link.

The relative interaction abundances for plants, \(\sigma_i\), has a Dirichlet prior that accepts a vector of concentration parameters, alpha_sigma, with a length equal to the number of plant species (rows) in the interaction matrix. The default, \(\mathrm{Dir}(1, ..., 1)\), implies a uniform distribution of proportions. Increasing some of the concentration parameters will shift the probability mass to favor higher relative interaction abundances for the corresponding plants. Increasing all concentration parameters while maintaining their relative sizes reduces the variance.

The relative interaction abundances for animals, \(\tau_j\), has a Dirichlet prior that accepts a vector of concentration parameters, alpha_tau, with a length equal to the number of animal species (columns) in the interaction matrix. The default, \(\mathrm{Dir}(1, ..., 1)\), implies a uniform distribution of proportions. Increasing some of the concentration parameters will shift the probability mass to favor higher relative interaction abundances for the corresponding animals Increasing all concentration parameters while maintaining their relative sizes reduces the variance.

At this time, prior choices have not been implemented for sampling effort, \(C_i\), when plant_effort = "param".

Examples

if (FALSE) { # interactive()
data(web)
dt <- prepare_data(mat = web, plant_effort = rep(20, nrow(web)))
# fit the model used by Young et al. 2021
fit <- fit_model(dt, plant_effort = "param", animal_pref = "fixed",
                 plant_abun = "param", animal_abun = "param")
# changing options and priors
fit2 <- fit_model(dt, plant_effort = "data", animal_pref = "varying",
                  plant_abun = "param", animal_abun = "param",
                  beta = rep(0.01, 21),
                  alpha_sigma = c(2.5, 3, 4, 3, 1, 1, 5, 2),
                  alpha_tau = 2)
}