Fit bipartite network model

Fit bipartite network model

Usage

fit_model(
  data = NULL,
  plant_effort = c("param", "data"),
  animal_pref = c("fixed", "varying"),
  plant_abun = c("param", "data"),
  animal_abun = c("param", "data"),
  beta = 0.01,
  alpha_sigma = 1,
  alpha_tau = 1,
  ...
)

Arguments

data

A named list containing the required data, as obtained from prepare_data().

plant_effort

How to model the sampling effort allocated to each plant. If unknown, param (default) estimates a single parameter during the modelling step that applies to all plants, \(C\). Alternatively, "data" allows measured values of sampling effort to be supplied in data, \(C_i\).

animal_pref

How to model the increase in visits of animals to plants with which they share a link relative to unlinked plants. fixed (default), fits a single preference parameter for all animals during the modelling step, \(r\). varying fits a separate preference parameter for each animal type, \(r_i\).

plant_abun

How the model treats the abundances of interacting plants. param (default) treats these abundances as a parameter that the model estimates, \(sigma_i\). data treats these abundances as measured and requires the user to supply the absolute abundances of each plant in data.

animal_abun

How the model treats the abundances of interacting animals. param (default) treats these abundances as a parameter that the model estimates, \(sigma_i\). data treats these abundances as measured and requires the user to supply the absolute abundances of each animal in data.

beta

Rate parameter of exponential prior on the preference parameter or parameters \(r\) or \(r_i\). The default, 0.01, is suitable for small counts and should be increased for larger counts. Use plot_prior to visualize.

alpha_sigma

Concentration parameters of Dirichlet prior on the abundances of interacting plants, \(sigma_i\), when plant_abun = "param". Ignored otherwise. Defaults to 1.

alpha_tau

Concentration parameters of Dirichlet prior on the abundances of interacting plants, \(tau_i\), when animal_abun = "param". Ignored otherwise. Defaults to 1.

...

Further arguments passed to cmdstanr::sample(), like iter_warmup, iter_sampling, or thin, among others.

Value

A fitted model (cmdstanr::CmdStanMCMC() object).

Details

This function allows fits generalized variations of the bipartite network structure model of Young et al. 2021 doi:10.1038/s41467-021-24149-x.

Examples

if (FALSE) { # interactive()
data(web)
dt <- prepare_data(mat = web, plant_effort = rep(20, nrow(web)))
# fit the model used by Young et al. 2021
fit <- fit_model(dt, plant_effort = "param", animal_pref = "fixed",
                 plant_abun = "param", animal_abun = "param")
}